task3/user_service/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
pyjwt>=2.1.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0